+++ example/bloc/main.dart
@@ -2,9 +2,11 @@ import 'provider.dart';
 import 'counter.dart';
 import 'package:flutter/material.dart';
 
-void main() => runApp(MyApp());
+void main() => runApp(const MyApp());
 
 class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
@@ -15,22 +17,22 @@ class MyApp extends StatelessWidget {
         theme: ThemeData(
           primarySwatch: Colors.blue,
         ),
-        home: MyHomePage(title: 'Flutter Demo Home Page'),
+        home: const MyHomePage(title: 'Flutter Demo Home Page'),
       ),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key? key, this.title}) : super(key: key);
+  const MyHomePage({Key? key, this.title}) : super(key: key);
 
   final String? title;
 
   @override
-  _MyHomePageState createState() => _MyHomePageState();
+  MyHomePageState createState() => MyHomePageState();
 }
 
-class _MyHomePageState extends State<MyHomePage> {
+class MyHomePageState extends State<MyHomePage> {
   @override
   Widget build(BuildContext context) {
     final CounterBloc bloc = BlocProvider.of(context);
@@ -44,7 +46,7 @@ class _MyHomePageState extends State<MyHomePage> {
           bloc.incrementCounter();
         },
         tooltip: 'Increment',
-        child: Icon(Icons.add),
+        child: const Icon(Icons.add),
       ),
     );
   }
@@ -57,8 +59,8 @@ class _MyHomePageState extends State<MyHomePage> {
       builder: (BuildContext context, snapshot) {
         return Center(
           child: Text(
-            "Clicked " + snapshot.data.toString() + " times!",
-            style: TextStyle(fontSize: 18),
+            "Clicked ${snapshot.data} times!",
+            style: const TextStyle(fontSize: 18),
           ),
         );
       },