+++ example/lib/main.dart
@@ -2,6 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:iconnect/iconnect.dart';
 import 'package:example/app/app.dart';
 
+import 'app/init_store.dart';
+
 void main() {
-  runApp(provider(const MyApp()));
+  runApp(provider(const MyApp(), initState));
 }